% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shade.R
\name{add_shade}
\alias{add_shade}
\title{Highlight a wavelength range on a spectral plot}
\usage{
add_shade(
  obj,
  type = "em",
  fluo = NULL,
  min = NULL,
  max = NULL,
  range = NULL,
  fill = "white",
  alpha = 0.3,
  annotate_pct = F
)
}
\arguments{
\item{obj}{A ggplot object returned by `plot_spectra()`.}

\item{type}{Spectrum type to shade: `"em"` or `"ex"`.}

\item{fluo}{Optional fluorochrome name. Required when the plot contains
multiple fluorochromes.}

\item{min}{Minimum wavelength (nm).}

\item{max}{Maximum wavelength (nm).}

\item{range}{Character string specifying a filter as `"center/width"`
(e.g. `"530/30"`). Overrides `min` and `max`.}

\item{fill}{Fill colour of the shaded region.}

\item{alpha}{Ribbon transparency.}

\item{annotate_pct}{Logical; if `TRUE`, annotate the plot with the
percentage of the emission spectrum within the selected range.}
}
\value{
A ggplot object with the shaded wavelength region added.
}
\description{
Adds a shaded ribbon to a spectral plot over a specified wavelength range.
The range can be supplied either as minimum and maximum wavelengths or
using filter notation such as `"530/30"`. Optionally annotates the plot
with the percentage of the emission spectrum contained within the selected
range.
}
\examples{
library(muchofluo)
data <- get_spectral_data(fluos = c("FITC", "PE-cy7"))
pp <- plot_spectra(data$bd)[[1]]
add_shade(pp,
          fluo = "FITC",
          range = "515/50",
          annotate_pct = TRUE)
add_shade(pp,
          fluo = "FITC",
          annotate_pct = TRUE)
}
\seealso{
[filter_pct()], [plot_spectra()]
}
