% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_filter.R
\name{geom_filter}
\alias{geom_filter}
\title{Add filter to spectral plot}
\usage{
geom_filter(min = NULL, max = NULL, range = NULL, fill = "grey90", alpha = 0.2)
}
\arguments{
\item{min}{lower limit}

\item{max}{upper limit}

\item{range}{range in form of 530/30, center/width}

\item{fill}{fill color}

\item{alpha}{opacity from 0 to 1}
}
\value{
geom rect
}
\description{
Add filter to spectral plot
}
\examples{
plot_spectra(get_spectral_data(c("PE", "FITC"))[[1]])[[1]] +
  geom_filter(range = "515/30")
}
