% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fluo.R
\name{query_fluo}
\alias{query_fluo}
\title{Title}
\usage{
query_fluo(
  query,
  multimatch = c("best", "all"),
  ref = system.file("extdata", "fluo_lookup.tsv", package = "muchofluo")
)
}
\arguments{
\item{ref}{}
}
\description{
Title
}
