% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectral_data.R
\name{get_spectral_data}
\alias{get_spectral_data}
\title{Title}
\usage{
get_spectral_data(
  fluos,
  spectra_df = vroom::vroom(system.file("extdata", "spectra.tsv.gz", package =
    "muchofluo"), show_col_types = F, progress = F),
  maxima_df = vroom::vroom(system.file("extdata", "em_ex_maxima.tsv", package =
    "muchofluo"), show_col_types = F, progress = F),
  progress = F
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
\examples{
out <- get_spectral_data(c("A647", "FITC"))
}
