% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spectra_plots.R
\name{make_spectra_plots}
\alias{make_spectra_plots}
\title{Make ggplots of emission and excitation spectra}
\usage{
make_spectra_plots(
  spectra_df = system.file("extdata", "spectra.tsv.gz", package = "muchofluo"),
  maxima_df = system.file("extdata", "em_ex_maxima.tsv", package = "muchofluo"),
  path = file.path(getwd(), "spectra_plots"),
  format = "pdf",
  width = 5,
  height = 3.5
)
}
\arguments{
\item{spectra_df}{data frame of path to file on disk with spectral data}

\item{maxima_df}{data frame of path to file on disk with maxima/peaks of
emission and excitation}

\item{path}{path to folder where to save plots to}

\item{format}{format to save, see ?ggsave for possible formats}

\item{width}{width of plots}

\item{height}{height of plots}
}
\value{
nothing, files on disk created
}
\description{
Make ggplots of emission and excitation spectra
}
\examples{
\dontrun{
# run with all values default
make_spectra_plots()
}
}
