% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Title}
\usage{
plot_spectra(spec_data)
}
\arguments{
\item{spec_data}{}
}
\description{
Title
}
\examples{
out <- get_spectral_data(c("A647", "FITC"))
plots <- plot_spectra(out)
}
