% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fluo.R
\name{query_fluo}
\alias{query_fluo}
\title{Query the}
\usage{
query_fluo(
  query,
  multimatch = c("best", "all"),
  bestmatch_cutoff = 0.5,
  ref = system.file("extdata", "fluo_lookup.tsv", package = "muchofluo")
)
}
\arguments{
\item{ref}{}
}
\description{
Query the
}
\examples{
query <- c("abc", "pr", "0", "pecy5", "pecy")
multimatch <- "best"
}
